Recursive error corrected
An error had been detected when the recursive process was launched with two directories with the same files. The problem was that there was no way to differentiate two files belonging to different directories by their name. To correct this, now the field "file_name" of a Node is made up of the full path and the filename itself (without extension)